requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9
requests==2.31.0
python-docx==1.1.0
markdownify==0.12.1
Pillow==10.4.0